add paths-table.  disable right panel Block tab unless options=blockTab
add paths-table.{hbs,js}
draw/block-adj.js : split functions out to form paths-api.js, for use of pathsResultTypes in paths-table.
models/block-adj.js : add adjacentTo().
mapview.hbs : check parseOptions.blockTab around button Block.  Add Paths tab, use panel/paths-table.
 ddee626 5835 Feb 12 15:21 frontend/app/components/panel/paths-table.js
 7edff76 1276 Feb 12 15:35 frontend/app/templates/components/panel/paths-table.hbs
 6cc3d83 8139 Feb 12 13:08 frontend/app/utils/paths-api.js
