+++ frontend/app/styles/app.scss
@@ -195,6 +195,11 @@ div#logo {
 {
 overflow-x: scroll;
 }
+
+table.contextual-data-table td {
+  padding-right: 1em;
+}
+
 /* -------------------------------------------------------------------------- */
 
 svg {
